Split browser HTML snapshot builder
