+++ docs/DECISIONS.md
@@ -281,3 +281,4 @@
 - Tool artifact state and preview derivation live in `QuillCodeToolArtifactSurface.swift`, not in `QuillCodeToolCardSurface.swift`. Tool cards own status, actions, density, review substates, and card-level artifact grouping; artifact kind detection, image/document preview metadata, href/detail labels, and bounded text preview extraction evolve beside the artifact DTO so transcript projection, activity, native cards, and static HTML share one artifact contract.
 - Sidebar thread list derivation lives in `QuillCodeSidebarThreadListBuilder.swift`, not in the aggregate `SidebarSurface` DTO. `SidebarSurface` keeps the stable public API consumed by native SwiftUI, static HTML, and search dialogs, while filtering, pinned/recent/archived partitioning, and date-bucket sectioning stay in the focused helper.
 - Native sidebar command payloads are built through `QuillCodeSidebarCommandAdapter.swift`. SwiftUI row and bulk-action views should not construct `WorkspaceCommandSurface` values inline; selection toggles and bulk actions share the same adapter so command IDs, titles, categories, and enabled state stay consistent.
+- Static HTML browser snapshots live in `BrowserHTMLSnapshotBuilder`; `BrowserInspector` remains the URL/file/fetched-page adapter and delegates title/count/outline/snippet extraction to the focused builder.
